This PR brings a few features:
- usage of `Format` boxes to provide the same formatting as `deriving show` _and_ abide by `Format`’s rule that all content should be output into some box,
- the possibility to require the output to be truncated,
- a couple of pretty-printers that were missing,
- a couple of typo fixes.

The possibility to truncate would address #308 if all pretty-printers in `STM` and `Lin` were always using `Util.Pp` (this is #343).