@jmid 2023-08-21
LGTM, thanks!
I think an environment variable is the right solution, to allow for both a "shortened print" and a "full one" without having to recompile.

Should we do something similar for `bytes`?
Would it make sense to add unit/expect tests to ensure that it works as intended and protect ourselves against future breakage? :thinking: 
Finally (for another PR) I agree that we should seriously consider switching Lin+STM to use these for consistent 
argument and result printing.