@shym 2023-08-22
I’ve added an internal test for the pretty-printers, with variants for the truncation.

> Should we do something similar for bytes?
I don’t see what you have in mind, `Util.Pp` contains a `pp_bytes` pretty-printer.