@jmid 2023-08-23
> I don’t see what you have in mind, `Util.Pp` contains a `pp_bytes` pretty-printer.

My bad, I was too fast reading the cut off code as only applying to strings and not bytes :facepalm: 
Since this is located in `to_show` I assume the cut off applies to all to-string-coercions (string, bytes, list, ...) passing through that function, right?