@shym 2023-08-23
> The added tests look great BTW, and the expect outputs nicely answer the above question 😅

:sweat_smile:

> I guess having it print the final closing quote or square bracket like the OCaml toplevel will require some form of type-directed approach?

I think so. At the level where the truncating mechanism is located in that PR, it sees only the stream of bytes to output, without structure. I wondered whether the boxing or tagging mechanisms in formatters could be retargeted for that kind of purposes, but I didn’t find how.