@shym 2023-08-23
Reviewing the CI logs I realise (finally!) that (obviously, as it is one of the main goal of `Format`) the result produced by `to_show` can be a multiline string while `print_triple_vertical` assumes it is not the case (see this [example](https://github.com/ocaml-multicore/multicoretests/actions/runs/5942185755/job/16114588868?pr=379#step:18:239) for instance).
(It must have been the case already with the pretty-printers generated by `ppx_deriving.show` since we have something really similar, hidden by the fact that long lines are not readable anyway)

A simple way to fix pretty-printing with `print_triple_vertical` is to set the formatter margin to `max_int`.
But now I wonder: do we want `to_show` functions to always generate one-liners? Or should we add the margin as an (optional?) argument to override the default 78? :thinking: