@jmid 2023-08-23
First of all, let me state that `print_triple_vertical` is an ugly piece of code that I'm not proud of! :sweat_smile: 
(I don't know whether it can be written more cleanly with Formatters).
I agree that it assumes 1-line `cmd` strings when stitching them together.

Code aside, too long outputs (strings, bytes, lists, ...) mangle the upside down Y-shape printed, which motivated the request for abbreviating print functions in #308

In CI logs, I'd rather skim over, e.g., an abbreviated string in an comprehendable Y-shape rather than see a long string in full detail. That suggests we should set the environment variable in the CI workflows.
I suspect other end-users would prefer such a default setup too :thinking: 

It would however be nice to be able to print the full string some way (locally and in the CI), say to create a standalone reproducer. That should however be possible by resetting the `MCTUTILS_TRUNCATE` environment variable.

To keep things nice and without wrapped lines in the former scenario, setting the formatter margin to `max_int` sounds reasonable.

There may be cases I'm overlooking, say record types with lists in them that could warrant multi-line formatted strings. So far, such types haven't occurred in the interfaces we have tested :shrug: