@jmid 2023-08-25
CI summary for the latest run 64c1902
- Cygwin trunk part 2 failed to trigger `STM Sys test parallel`
  before also timing out during a 253.7s `Lin CList int64 test with Thread`
  after a 2403.4s `STM Sys test parallel`, 1071.1s `STM Weak test parallel`, 3267.4s `STM Weak HashSet test parallel`, 5028.3s `Lin ref int64 test with Thread` and 1424.6s `Lin CList int test with Thread` #384 
- Linux ppc64le failed with an `STM Bytes test sequential` segfault and an `STM Float Array test sequential` model difference #380

Neither of these are related to the PR, so I'll go ahead and merge