@jmid 2023-08-25
Actually, it would be nice with a CHANGES entry to document changes in the user-facing Util.Pp module.
Could I convince you to add one @shym?