@shym 2023-08-25
Indeed!
Should I add also a default setting to `MCTUTILS_TRUNCATE` to 50 I’ve tested locally? As `Util.Pp` are used only for STM commands, it won’t bring the hoped truncation for the worse offenders (`Dynlink`, say) just yet, though.