@jmid 2023-08-25
> Should I add also a default setting to `MCTUTILS_TRUNCATE` to 50 I’ve tested locally? 

Yep, please do! :pray: