@jmid 2023-08-25
CI summary for 13d87f8
- Win bytecode trunk/5.2 timeout during `threadomain` #203
- Cygwin trunk/5.2 timeout during 1343.7s `Lin CList int test with Thread` after 2276.9s `STM Sys test parallel`, 10291.7s `Lin ref int64 test with Thread` #384
- Linux ppc64le trunk/5.2 failure with a `STM Array test sequential` `ABRT` and a `STM Float Array test sequential` `SEGV` #380

None of these are related to the PR, so I'll merge (and I mean it this time!)